requirements.txt
numpy==1.24.3
pandas==1.5.3
backtrader==1.9.74.123
tensorflow==2.12.0
quickfix==1.15.2
pyyaml==6.0
websockets==11.0.3
